refactor: Move sink/tablesink to sink/type/tablesink
