+++ internal/pkg/service/stream/sink/type/tablesink/config.go
